requirements.txt
tree-sitter>=0.22.3
tree-sitter-python>=0.21.0
tree-sitter-javascript>=0.21.4
PyYAML>=6.0.1
rich>=13.7.1
